more gcode parser work.
git-svn-id: https://reprap.svn.sourceforge.net/svnroot/reprap/trunk/users/hoeken/arduino/firmware@1340 cb376a5e-1013-0410-a455-b6b1f9ac8223

+++ GCode_Interpreter/init.pde
@@ -1,7 +1,7 @@
 
 // our point structure to make things nice.
 struct Point {
-	int x;
-	int y;
- 	int z;
+	long x;
+	long y;
+ 	long z;
 };
